+++ cruise-control-client/cruisecontrolclient/client/Endpoint.py
@@ -171,6 +171,7 @@ class AddBrokerEndpoint(AbstractEndpoint):
         CCParameter.ExcludedTopicsParameter,
         CCParameter.GoalsParameter,
         CCParameter.JSONParameter,
+        CCParameter.ReasonForRequestParameter,
         CCParameter.ReviewIDParameter,
         CCParameter.ReplicaMovementStrategiesParameter,
         CCParameter.SkipHardGoalCheckParameter,
@@ -239,6 +240,7 @@ class DemoteBrokerEndpoint(AbstractEndpoint):
         CCParameter.ExcludeFollowerDemotionParameter,
         CCParameter.ExcludeRecentlyDemotedBrokersParameter,
         CCParameter.JSONParameter,
+        CCParameter.ReasonForRequestParameter,
         CCParameter.ReplicaMovementStrategiesParameter,
         CCParameter.ReviewIDParameter,
         CCParameter.SkipURPDemotionParameter,
@@ -261,6 +263,7 @@ class FixOfflineReplicasEndpoint(AbstractEndpoint):
     http_method = "POST"
     can_execute_proposal = True
     available_Parameters = (
+        CCParameter.ReasonForRequestParameter,
         CCParameter.ReviewIDParameter,
     )
     argparse_properties = {
@@ -333,7 +336,7 @@ class PauseSamplingEndpoint(AbstractEndpoint):
     can_execute_proposal = False
     available_Parameters = (
         CCParameter.JSONParameter,
-        CCParameter.ReasonParameter,
+        CCParameter.ReasonForPauseParameter,
         CCParameter.ReviewIDParameter,
     )
     argparse_properties = {
@@ -382,6 +385,7 @@ class RebalanceEndpoint(AbstractEndpoint):
         CCParameter.GoalsParameter,
         CCParameter.IgnoreProposalCacheParameter,
         CCParameter.JSONParameter,
+        CCParameter.ReasonForRequestParameter,
         CCParameter.ReplicaMovementStrategiesParameter,
         CCParameter.ReviewIDParameter,
         CCParameter.SkipHardGoalCheckParameter,
@@ -414,6 +418,7 @@ class RemoveBrokerEndpoint(AbstractEndpoint):
         CCParameter.ExcludedTopicsParameter,
         CCParameter.GoalsParameter,
         CCParameter.JSONParameter,
+        CCParameter.ReasonForRequestParameter,
         CCParameter.ReplicaMovementStrategiesParameter,
         CCParameter.ReviewIDParameter,
         CCParameter.SkipHardGoalCheckParameter,
@@ -438,7 +443,7 @@ class ResumeSamplingEndpoint(AbstractEndpoint):
     can_execute_proposal = False
     available_Parameters = {
         CCParameter.JSONParameter,
-        CCParameter.ReasonParameter,
+        CCParameter.ReasonForPauseParameter,
         CCParameter.ReviewIDParameter,
     }
     argparse_properties = {
@@ -456,7 +461,7 @@ class ReviewEndpoint(AbstractEndpoint):
         CCParameter.ApproveParameter,
         CCParameter.DiscardParameter,
         CCParameter.JSONParameter,
-        CCParameter.ReasonParameter
+        CCParameter.ReasonForReviewParameter
     )
     argparse_properties = {
         'args': (name,),
@@ -533,6 +538,7 @@ class TopicConfigurationEndpoint(AbstractEndpoint):
         CCParameter.JSONParameter,
         CCParameter.SkipHardGoalCheckParameter,
         CCParameter.SkipRackAwarenessCheckParameter,
+        CCParameter.ReasonForRequestParameter,
         CCParameter.ReplicaMovementStrategiesParameter,
         CCParameter.ReplicationFactorParameter,
         CCParameter.ReviewIDParameter,
